remove SettingDataSource
