+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/di/dataProviderModule.kt
@@ -3,21 +3,12 @@ package net.thechance.mena.core_chat.data.di
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import com.bilalazzam.contacts_provider.ContactsProvider
-import net.thechance.mena.core_chat.data.contacts.source.device.DeviceContactsDataSource
-import net.thechance.mena.core_chat.data.contacts.source.device.SettingDataSource
-import net.thechance.mena.core_chat.data.contacts.source.remote.DummyContactsDataSource
-import org.koin.core.qualifier.qualifier
 import org.koin.core.scope.Scope
 import org.koin.dsl.module
 
-const val SETTINGS_DATA_STORE = "settingsDataStore"
-
 internal val dataProviderModule = module {
-
-internal val dataSourceModule = module {
     single { createContactsProvider() }
-    single(qualifier(name = SETTINGS_DATA_STORE)) { createSettingsDataStore() }
-    single { SettingDataSource(get(qualifier(name = SETTINGS_DATA_STORE))) }
+    single { createSettingsDataStore() }
 }
 
 expect fun Scope.createContactsProvider(): ContactsProvider
